Workaround incorrect telemetry capability application
In the case where an object support telemetry but does not support
the specific type of telemetry request i.e. a summary widget has
request for lad but not for historical, an error will be thrown.

This works around that issue by allowing the provider to return an
empty response when the request is not a lad query.  This should be
refactored at a later date.